chore: move meta tag and open browsers utils to core
