+++ packages/core/static/openChrome.applescript
