--- packages/shared/compiled/open/package.json
@@ -1,1 +1,0 @@
-{"name":"open","author":{"name":"Sindre Sorhus","email":"sindresorhus@gmail.com","url":"https://sindresorhus.com"},"version":"8.4.0","funding":"https://github.com/sponsors/sindresorhus","license":"MIT"}
