+++ packages/shared/src/index.ts
@@ -6,7 +6,6 @@ export * from './pluginStore';
 export * from './createContext';
 export * from './utils';
 export * from './fs';
-export * from './generateMetaTags';
 export * from './getBrowserslist';
 export * from './getCssSupport';
 export * from './logger';
@@ -23,7 +22,6 @@ export * from './prodServer';
 export * from './chain';
 export * from './schema';
 export * from './plugins';
-export * from './openBrowser';
 export * from './apply';
 export * from './fallback';
 export * from './getLoaderOptions';
