+++ packages/shared/tests/config.test.ts
@@ -2,7 +2,6 @@ import webpack from 'webpack';
 import {
   getExtensions,
   stringifyConfig,
-  getMetaTags,
   pickRsbuildConfig,
 } from '../src/config';
 
@@ -95,26 +94,6 @@ describe('stringifyConfig', () => {
     };
 
     expect(await stringifyConfig(rsbuildConfig)).toMatchSnapshot();
-  });
-
-  it('shold meta generate correctly', async () => {
-    const rsbuildConfig = {
-      html: {
-        meta: { description: 'This is basic meta', bar: 'bar', foo: 'foo' },
-        metaByEntries: {
-          entry1: {
-            description: 'This is about page',
-          },
-        },
-      },
-      output: {} as any,
-    };
-
-    const defaultEntry = await getMetaTags('', rsbuildConfig);
-    expect(defaultEntry).toMatchSnapshot();
-
-    const entry1 = await getMetaTags('entry1', rsbuildConfig);
-    expect(entry1).toMatchSnapshot();
   });
 });
