+++ packages/shared/tests/tsconfig.json
@@ -4,5 +4,5 @@
     "baseUrl": "./",
     "types": ["vitest/globals"]
   },
-  "include": ["./**/*.ts"]
+  "include": ["./**/*.ts", "../../core/tests/plugins/openBrowser.test.ts"]
 }
